Add a Run workflow page
